Save order as a result of the form submission
Before this change, a javascript listener would generate the
order when the submit button was clicked, without any associated
validations.

This change puts the order contents as a form hidden field and
executes the order request as a result of the POST request occurred
when the form is submitted.

Note: This approach has the added advantage of having the order
be a snapshot of what the cart was at the moment of navigating
to the "confirm e-mail" screen.

Pending action items:
- When the order submits successfuly, clear the localstorage.
  This is a known limitation of this change.
- In the confirmation screen, we should also provide the user
  with a summary of the order like the one in the previous
  screen, but obtaining product data from the server, instead
  of local storage (prices could be out of date, etc).

Issue: #14